pyproject.toml
[project]
name = "langchain-trials"
version = "0.1.0"
description = ""
authors = [
    {name = "RevanKumarD",email = "revan.dhana@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain-ollama (>=0.2.3,<0.3.0)",
    "ollama (>=0.4.7,<0.5.0)",
    "langchain-core (>=0.3.33,<0.4.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)",
    "langgraph (>=0.2.69,<0.3.0)",
    "langsmith (>=0.3.4,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langchain-openai (>=0.3.3,<0.4.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
